Cloud Integrity Check Pre-Integration
Discovered a potential bug where cloud data can dissappear or become altered - this pre-integration outlines the bare beginnings of an algorithm which will use the local database to validate or repair the cloud environment if necessary. This is similar to the file creation error checking during which new events dynamically plug info into any existing gaps if they exist. The difference is that previously finished events that have been deleted or altered have no way to be restored automatically.