Cloud Integrity System Finished
This is actually beautiful. It has also given me an idea; if a folder that is deleted can be re-generated from the database, then it would therefore be possible to update already-created events via bot commands. The bot can update the db, delete the folder, and re-generate the data as if it were new

This has flaws, but could potentially be implemented later-on